Add build script
Adds script to build the lambda function bundle. In addition this script can be added as a pre-commit hook and will only activate when the committed changes include changes to `index.py` or the requirements file for the function. This allows the developer to be sure the function bundle is always updated.